I now can render text... i keep track of score and health... there is no end state or start menu at the moment

+++ ShapesOfCombat/ShapesOfCombat/ShapesOfCombat/main.cpp
@@ -305,7 +305,7 @@ int WINAPI wWinMain(HINSTANCE hInstance, HINSTANCE prevInstance, LPWSTR cmd, int
 
 			clock_t tend = clock();
 
-			float deltaTime = (tend - oldTime) / (double)CLOCKS_PER_SEC;
+			float deltaTime = (float)((tend - oldTime) / (double)CLOCKS_PER_SEC);
 
 			oldTime = clock();
